requirements.txt
Flask==3.0.3
keras==3.0.5
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.16.1